test/PlaceInput.test.tsx
// PlaceInput.test.tsx

import React from 'react';
import { render, fireEvent, screen } from '@testing-library/react';
import PlaceInput from '../src/controls/place/PlaceInput'; // Replace with the correct import path

describe('PlaceInput component', () => {
  test('renders without errors', () => {
    render(<PlaceInput />);
    // Add your assertions here, e.g., checking if certain elements are in the document
    expect(screen.getByText('Move Your Robot!')).toBeInTheDocument();
  });

  test('handles input change', () => {
    render(<PlaceInput />);
    const inputElement = screen.getByPlaceholderText('Place x,y North');
    fireEvent.change(inputElement, { target: { value: 'place 1,2 north' } });
    expect(inputElement).toHaveValue('place 1,2 north');
  });

  test('submits the form', () => {
    const mockSubmit = jest.fn();
    render(<PlaceInput />);
    const submitButton = screen.getByText('Submit');
    fireEvent.click(submitButton);
    expect(mockSubmit).toHaveBeenCalledTimes(1); // Ensure that your submit function was called
  });

  // Add more test cases as needed
});
